Merge ModuleReports with different artefact types
Under sbt-coursier it's possible to have multiple ModuleInfo for the
same (groupId, artifactId), where there is no difference between them,
or the difference is one has artifacts of type "jar" and the other of
type "bundle".

For this last case we 'upgrade' "jar" to "bundle".

Additionally the distinct ModuleReports can have different callers, so
ignore that too.

This is achieved by discarding most of ModuleReport, and then merging
the remaining info, carefully.